feat: global.css 수정
