+++ index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html>
+<!doctype html>
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
@@ -9,7 +9,7 @@
     <title>Vite + React + TS</title>
   </head>
   <body>
-    <div id="root" class="bg-900 m-auto flex min-h-screen w-[480px] flex-col"></div>
+    <div id="root" class="bg-c900 m-auto flex min-h-screen w-[480px] flex-col"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
