+++ src/router.tsx
@@ -8,6 +8,7 @@ import PlaylistCollectionPage from './pages/PlaylistCollection/PlaylistCollectio
 import ProfilePage from './pages/Profile/ProfilePage'
 import LoginPage from './pages/Login/LoginPage'
 import SignupPage from './pages/Signup/SignupPage'
+import DesignSystem from './pages/DesignSystem/DesignSystem'
 
 const routes: RouteObject[] = [
   {
@@ -47,6 +48,10 @@ const routes: RouteObject[] = [
   {
     path: '/signup',
     element: <SignupPage />,
+  },
+  {
+    path: '/designsystem',
+    element: <DesignSystem />,
   },
 ]
