+++ src/shared/components/ui/input.tsx
@@ -8,9 +8,9 @@ function Input({ className, type, ...props }: React.ComponentProps<'input'>) {
       type={type}
       data-slot="input"
       className={cn(
-        'bg-300 text-900 placeholder:text-500 text-textR flex h-12 w-full rounded-xl px-4 py-3',
+        'bg-c300 text-c900 placeholder:text-c500 text-textR flex h-12 w-full rounded-xl px-4 py-3',
         'transition-colors outline-none',
-        'focus:bg-50',
+        'focus:bg-c50',
         'disabled:pointer-events-none disabled:cursor-not-allowed disabled:opacity-50',
         className,
       )}
